# docs/swagger.yaml
basePath: /
definitions:
...
  api_handler.ErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/api_handler.FieldError'
        type: array
      message:
        example: Invalid request
        type: string
      status:
        example: 400
        type: integer
    type: object
...
  api_handler.FieldError:
    properties:
      field:
        example: email
        type: string
      message:
        example: Email is required
        type: string
    type: object
...
  github_com_edsonmubezi_myapp_internal_orgsettings.OrganizationSettings:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      loanModuleEnabled:
        type: boolean
      organizationId:
        type: integer
      organizationType:
        enum:
        - single_company
        - multiple_company
        - multi_branch
        type: string
      selfServiceEnabled:
        type: boolean
      updatedAt:
        type: string
    required:
    - organizationType
    type: object
  github_com_edsonmubezi_myapp_internal_orgsettings.OrganizationSettingsCreateInput:
    properties:
      loanModuleEnabled:
        type: boolean
      organizationId:
        type: integer
      organizationType:
        enum:
        - single_company
        - multiple_company
        - multi_branch
        type: string
      selfServiceEnabled:
        type: boolean
    required:
    - organizationId
    - organizationType
    type: object
  github_com_edsonmubezi_myapp_internal_orgsettings.OrganizationSettingsUpdateInput:
    properties:
      id:
        type: integer
      loanModuleEnabled:
        type: boolean
      organizationType:
        enum:
        - single_company
        - multiple_company
        - multi_branch
        type: string
      selfServiceEnabled:
        type: boolean
    required:
    - id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ekklesiahr.com
    name: EkklesiaHR API Support
    url: https://github.com/edsonmubezi/ekklesiahr
  description: Production-ready HR Management System API with comprehensive employee,
    payroll, leave management, and organization administration. Supports multi-tenant
    architecture with role-based access control.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: EkklesiaHR API
  version: 1.0.1
paths:
...
  /api/v1/super/org-settings:
    post:
      consumes:
      - application/json
      description: Creates core settings for an organization (Super Admin only)
      parameters:
      - description: Organization settings payload
        in: body
        name: orgSettings
        required: true
        schema:
          $ref: '#/definitions/github_com_edsonmubezi_myapp_internal_orgsettings.OrganizationSettingsCreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_edsonmubezi_myapp_internal_orgsettings.OrganizationSettings'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Organization Settings
      tags:
      - OrganizationSettings
    put:
      consumes:
      - application/json
      description: Updates organization settings (Super Admin only)
      parameters:
      - description: Organization settings update payload
        in: body
        name: orgSettings
        required: true
        schema:
          $ref: '#/definitions/github_com_edsonmubezi_myapp_internal_orgsettings.OrganizationSettingsUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_edsonmubezi_myapp_internal_orgsettings.OrganizationSettings'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Organization Settings
      tags:
      - OrganizationSettings
  /api/v1/super/org-settings/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves organization settings by its ID (Super Admin only)
      parameters:
      - description: Organization Settings ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_edsonmubezi_myapp_internal_orgsettings.OrganizationSettings'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Organization Settings By ID
      tags:
      - OrganizationSettings
  /api/v1/super/org-settings/organization/{organizationId}:
    get:
      consumes:
      - application/json
      description: Retrieves organization settings by organization ID (Super Admin
        only)
      parameters:
      - description: Organization ID
        in: path
        name: organizationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_edsonmubezi_myapp_internal_orgsettings.OrganizationSettings'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Organization Settings By Organization ID
      tags:
      - OrganizationSettings
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: 'Provide your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"